move new config stuff into the correct package
